refactor: 패키지 이름 di로 변경
