+++ app/src/main/java/com/sdhong/pokemonapp/datastore/module/DataStoreModule.kt
@@ -11,8 +11,8 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
-@InstallIn(SingletonComponent::class)
 @Module
+@InstallIn(SingletonComponent::class)
 object DataStoreModule {
 
     private const val TIME_DATASTORE_NAME = "TIME_PREFERENCES"
